api/sdlc-integration/requirements.txt
python-owasp-zap-v2.4
requests
